feat #206: Adaptando estrutura do módulo `location`
