+++ lib/src/app/modules/location/domain/entities/response/response_get_locations_near_user.dart
