+++ lib/src/app/modules/location/domain/entities/response/response_save_location.dart
@@ -1,4 +1,4 @@
-import 'package:is_it_safe_app/src/service/api/modules/location/response/response_location_review.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/response/response_location_review.dart';
 
 class ResponseSaveLocation {
   int? id;
