+++ lib/src/app/modules/location/domain/usecases/get_locations_by_id_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:is_it_safe_app/src/app/modules/location/error/safe_location_error.dart';
+import 'package:is_it_safe_app/src/app/modules/location/services/location_service_interface.dart';
 import 'package:is_it_safe_app/src/core/interfaces/safe_use_case.dart';
-import 'package:is_it_safe_app/src/domain/entity/location_entity.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/location_service_interface.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/location_entity.dart';
 import 'package:result_dart/result_dart.dart';
 
 class GetLocationsByIdUseCase extends SafeUseCase {
