+++ lib/src/app/modules/location/presenter/bloc/location_bloc.dart
@@ -5,17 +5,17 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:is_it_safe_app/generated/l10n.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/usecases/get_locations_by_id_use_case.dart';
 import 'package:is_it_safe_app/src/core/constants/string_constants.dart';
 import 'package:is_it_safe_app/src/core/enum/location_type_enum.dart';
 import 'package:is_it_safe_app/src/core/interfaces/safe_bloc.dart';
 import 'package:is_it_safe_app/src/core/state/safe_stream.dart';
 import 'package:is_it_safe_app/src/core/util/parse_enum.dart';
 import 'package:is_it_safe_app/src/core/util/validation_util.dart';
-import 'package:is_it_safe_app/src/domain/entity/location_entity.dart';
-import 'package:is_it_safe_app/src/domain/use_case/get_locations_by_id_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/location_entity.dart';
 import 'package:is_it_safe_app/src/domain/use_case/save_location_use_case.dart';
-import 'package:result_dart/result_dart.dart';
 import 'package:flutter_modular/flutter_modular.dart';
+import 'package:result_dart/result_dart.dart';
 
 import '../../../../../service/api/configuration/api_interceptors.dart';
 import '../../../../../service/api/error/error_exceptions.dart';
