+++ lib/src/app/modules/search/search_module.dart
@@ -1,11 +1,10 @@
 import 'package:flutter_modular/flutter_modular.dart';
+import 'package:is_it_safe_app/src/app/modules/location/services/location_service.dart';
 import '../auth/services/auth_service.dart';
 import '../../../domain/use_case/get_locations_by_name_use_case.dart';
 import 'presenter/bloc/search_bloc.dart';
 import 'presenter/pages/search_page.dart';
 import '../../../service/api/modules/search/search_service.dart';
-
-import '../../../service/api/modules/location/location_service.dart';
 
 class SearchModule extends Module {
   @override
