+++ lib/src/domain/use_case/save_location_use_case.dart
@@ -1,12 +1,12 @@
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/request/request_save_location.dart';
 import 'package:is_it_safe_app/src/app/modules/location/error/safe_location_error.dart';
+import 'package:is_it_safe_app/src/app/modules/location/services/location_service_interface.dart';
 import 'package:is_it_safe_app/src/core/interfaces/safe_use_case.dart';
-import 'package:is_it_safe_app/src/domain/entity/location_entity.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/location_service_interface.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/request/request_save_location.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/location_entity.dart';
 import 'package:result_dart/result_dart.dart';
 
 class SaveLocationUseCase extends SafeUseCase {
-  late final ILocationService service;
+  final ILocationService service;
 
   SaveLocationUseCase(this.service);
