+++ test/app/modules/location_module/mocks/location_mocks.dart
@@ -1,4 +1,4 @@
-import 'package:is_it_safe_app/src/service/api/modules/location/location_service.dart';
+import 'package:is_it_safe_app/src/app/modules/location/services/location_service.dart';
 import 'package:mocktail/mocktail.dart';
 
 class LocationServiceMock extends Mock implements LocationService {}
