+++ test/app/modules/location_module/mocks/mock_location_service.dart
@@ -1,12 +1,12 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:is_it_safe_app/src/service/api/modules/location/location_service_interface.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/response/response_save_location.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/response/response_get_locations_near_user.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/response/response_get_location_by_id.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/response/response_get_best_rated_places.dart';
-import 'package:is_it_safe_app/src/service/api/modules/location/request/request_save_location.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/request/request_save_location.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/response/response_get_best_rated_places.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/response/response_get_location_by_id.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/response/response_get_locations_near_user.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/entities/response/response_save_location.dart';
+import 'package:is_it_safe_app/src/app/modules/location/services/location_service_interface.dart';
 
 class MockLocationService implements ILocationService {
   @override
